+++ client/src/types/api.ts
@@ -1,4 +1,14 @@
-export interface IResponseInstallError {
+export interface IError {
+  type: string;
+  message: string;
+}
+
+export interface IResponseInstallError extends IError {
+  type: string;
+  message: string;
+}
+
+export interface IResponseError extends IError {
   type: string;
   message: string;
 }