+++ client/src/utils/defenders.ts
@@ -1,0 +1,3 @@
+export function instanceIResponseError(object: any): boolean {
+  return 'type' in object && 'message' in object;
+}