+++ fugitive-darkness-plugin/fugitive-darkness-plugin-vertx-git-provider/src/main/java/com/zer0s2m/fugitivedarkness/plugin/vertx/git/provider/impl/GitRepoProviderWebClientImpl.java
@@ -44,6 +44,7 @@ public class GitRepoProviderWebClientImpl implements GitRepoProviderWebClient {
         return client
                 .get(infoProvider.host(), infoProvider.path())
                 .expect(ResponsePredicate.SC_SUCCESS)
+                .expect(ResponsePredicate.JSON)
                 .putHeaders(MultiMap
                         .caseInsensitiveMultiMap()
                         .addAll(infoProvider.headers()))