+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/GitRepoUtils.java
@@ -1,5 +1,7 @@
 package com.zer0s2m.fugitivedarkness.provider.git;
 
+import com.zer0s2m.fugitivedarkness.common.Environment;
+
 public interface GitRepoUtils {
 
     String LINE_NUMBER_POINTER = "#L";
@@ -16,7 +18,13 @@ public interface GitRepoUtils {
             final ContainerGitRepoMeta gitRepoMeta,
             final String file,
             final String targetBranch) {
-        return gitRepoMeta.getLink(false) + "/tree/" + targetBranch + "/" + file;
+        return gitRepoMeta.getLink(false)
+                + "/"
+                + GitEquipment.REMOTE_TREE.value()
+                + "/"
+                + targetBranch
+                + "/"
+                + file;
     }
 
     /**
@@ -34,6 +42,19 @@ public interface GitRepoUtils {
             final String targetBranch,
             final int lineNumber) {
         return getLinkForFile(gitRepoMeta, file, targetBranch) + LINE_NUMBER_POINTER + lineNumber;
+    }
+
+    static String cleanRawFilePath(String path, String... parts) {
+        StringBuilder readyParts = new StringBuilder(Environment.ROOT_PATH_REPO);
+        if (!readyParts.substring(readyParts.length() - 1).equals('/')) {
+            readyParts.append('/');
+        }
+
+        for (String part : parts) {
+            readyParts.append(part).append('/');
+        }
+
+        return path.replace(readyParts.toString(), "");
     }
 
 }