+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/SearchEngineGrepStatistics.java
@@ -9,4 +9,11 @@ public interface SearchEngineGrepStatistics {
      */
     int getCountFiles();
 
+    /**
+     * Get the average file processing time.
+     *
+     * @return Average file processing time.
+     */
+    long getAverageFileProcessingTime();
+
 }