+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/impl/SearchIOFileCallable.java
@@ -45,13 +45,18 @@ class SearchIOFileCallable extends SearchInFileMatchFilterCallableAbstract<Conta
                 return null;
             }
 
-            return new ContainerInfoSearchFileGitRepo(
+            final String filename = GitRepoUtils.cleanRawFilePath(
                     file.toString(),
+                    containerGitRepoMeta.group(),
+                    containerGitRepoMeta.project());
+
+            return new ContainerInfoSearchFileGitRepo(
+                    filename,
                     FileSystemUtils.getExtensionFromRawStrFile(file.toString()),
                     GitRepoUtils.getLinkForFile(
                             containerGitRepoMeta,
-                            file.toString(),
-                            "master"),
+                            filename,
+                            currentBranch),
                     result);
         }
     }